# src/main/java/com/example/bankcards/exception/ErrorResponse.java
package com.example.bankcards.exception;

import java.time.Instant;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class ErrorResponse {
    private String code;
    private String message;
    private Instant timestamp;
}