# src/main/java/com/example/bankcards/exception/GlobalErrorResponse.java
package com.example.bankcards.exception;

import java.time.Instant;
import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class GlobalErrorResponse {
    private String code;
    private String message;
    private List<String> details;
    private Instant timestamp;
}